third_party/blink/web_tests/platform/generic/fast/forms/form-associated-element-expected.txt

Verify that only elements that are 'form associatable' get a form owner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS elements.length > 0 is true
Testing LABEL
PASS formOwner is "not defined"
Testing INPUT
PASS formOwner is "not defined"
Testing BUTTON
PASS formOwner is "defined"
Testing FIELDSET
PASS formOwner is "defined"
Testing INPUT
PASS formOwner is "defined"
Testing INPUT
PASS formOwner is "defined"
Testing LABEL
PASS formOwner is "not defined"
Testing INPUT
PASS formOwner is "not defined"
Testing OBJECT
PASS formOwner is "defined"
Testing SELECT
PASS formOwner is "defined"
Testing TEXTAREA
PASS formOwner is "defined"
Testing IMG
PASS formOwner is "defined"
Testing OPTION
PASS formOwner is "not defined"
Testing DIV
PASS formOwner is "not defined"
Testing A
PASS formOwner is "not defined"
Testing P
PASS formOwner is "not defined"
Testing INPUT
PASS formOwner is "defined"
Testing OBJECT
PASS formOwner is "defined"
Testing SELECT
PASS formOwner is "defined"
Testing PRE
PASS formOwner is "not defined"
Testing SPAN
PASS formOwner is "not defined"
Testing IMG
PASS formOwner is "not defined"


Tests for association-by-parser:
PASS form2['input1'] is defined.
PASS form2['image1'] is defined.
PASS form2['input-in-document'] is defined.
PASS form2['image-in-document'] is defined.
Detach a form, input1 and image1 from the document
PASS form2['input1'] is defined.
PASS form2['image1'] is defined.
PASS form2['input-in-document'] is undefined.
PASS form2['image-in-document'] is undefined.
Association-by-parser should not work for non-Document trees
PASS removed.querySelector('form')['image4'] is undefined.
PASS removed.querySelector('input').form is null
PASS removed.querySelector('object').form is null
PASS successfullyParsed is true

TEST COMPLETE
